refactor: member api, hooks
