+++ src/services/member/apis/index.ts
@@ -1,0 +1,15 @@
+export * from './certifyStudent';
+export * from './getMyInfo';
+export * from './getMyPortfolio';
+export * from './getProfileVisibility';
+export * from './getUserInfo';
+export * from './getUserPortfolio';
+export * from './getUserProfileVisibility';
+export * from './updateIsMajor';
+export * from './updateMyInfo';
+export * from './updateMyPortfolio';
+export * from './updateNickname';
+export * from './updateProfileVisibility';
+export * from './updateSsafyBasicInfo';
+export * from './updateTrack';
+export * from './validateNickname';
