+++ src/services/member/hooks/useCertifyStudent.ts
@@ -1,0 +1,9 @@
+import { useMutation } from '@tanstack/react-query';
+
+import { certifyStudent } from '~/services/member/apis';
+
+export const useCertifyStudent = () => {
+  return useMutation({
+    mutationFn: certifyStudent,
+  });
+};
