+++ src/services/member/hooks/useUpdateIsMajor.ts
@@ -1,0 +1,9 @@
+import { useMutation } from '@tanstack/react-query';
+
+import { updateIsMajor } from '~/services/member/apis';
+
+export const useUpdateIsMajor = () => {
+  return useMutation({
+    mutationFn: updateIsMajor,
+  });
+};
