+++ src/services/member/hooks/useUpdateNickname.ts
@@ -1,0 +1,9 @@
+import { useMutation } from '@tanstack/react-query';
+
+import { updateNickname } from '~/services/member/apis';
+
+export const useUpdateNickname = () => {
+  return useMutation({
+    mutationFn: updateNickname,
+  });
+};
