+++ src/services/member/hooks/useUpdateSsafyBasicInfo.ts
@@ -1,0 +1,9 @@
+import { useMutation } from '@tanstack/react-query';
+
+import { updateSsafyBasicInfo } from '~/services/member/apis';
+
+export const useUpdateSsafyBasicInfo = () => {
+  return useMutation({
+    mutationFn: updateSsafyBasicInfo,
+  });
+};
