+++ src/services/member/hooks/useUpdateTrack.ts
@@ -1,0 +1,9 @@
+import { useMutation } from '@tanstack/react-query';
+
+import { updateTrack } from '~/services/member/apis';
+
+export const useUpdateTrack = () => {
+  return useMutation({
+    mutationFn: updateTrack,
+  });
+};
