+++ src/services/member/hooks/useUserInfo.ts
@@ -1,0 +1,11 @@
+import { useQuery } from '@tanstack/react-query';
+
+import { queryKeys } from '~/react-query/common';
+import { getUserInfo } from '~/services/member/apis';
+
+export const useUserInfo = (id: number) => {
+  return useQuery({
+    queryKey: queryKeys.user.userInfo(id),
+    queryFn: () => getUserInfo(id),
+  });
+};
