+++ src/services/member/hooks/useValidateNickname.ts
@@ -1,0 +1,9 @@
+import { useMutation } from '@tanstack/react-query';
+
+import { validateNickname } from '~/services/member/apis';
+
+export const useValidateNickname = () => {
+  return useMutation({
+    mutationFn: validateNickname,
+  });
+};
